// budget-buddy-ui/src/app/services/models/budget-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BudgetRequest {
  amount: number;
  budgetCover?: string;
  description: string;
  endDate: string;
  id?: number;
  limitAmount: number;
  name: string;
  ownerId: number;
  startDate: string;
  targetAmount: number;
}